new: implement timeout edge parameter for nettrack

+++ src/nettrack.h
@@ -30,7 +30,7 @@ struct nt_vertex {
     struct nt_edges incoming_edges;
     // User defined actions on entry
     npc_match_fn *entry_fn;
-    unsigned timeout;   // if >0, number of seconds to keep an inactive state in here
+    int64_t timeout;   // if >0, number of seconds to keep an inactive state in here
     unsigned index_size;   // the index size (>=1)
     struct nt_states states[]; // the states currently waiting in this node (BEWARE: variable size!)
 };
